+++ src/main/java/com/vasa/scheduling/interfaces/web/TeamModifyController.java
@@ -75,7 +75,7 @@ public class TeamModifyController extends DefaultHandlerController{
 		String limit = request.getParameter("practiceLimit");
 		if(limit != null){
 			try{
-				team.setPracticeLimit(Integer.valueOf(limit));
+				team.setPracticeLimit(Double.valueOf(limit));
 			}catch(Exception e){
 				
 			}
@@ -86,7 +86,7 @@ public class TeamModifyController extends DefaultHandlerController{
 		limit = request.getParameter("weeklyPracticeLimit");
 		if(limit != null){
 			try{
-				team.setWeeklyPracticeLimit(Integer.valueOf(limit));
+				team.setWeeklyPracticeLimit(Double.valueOf(limit));
 			}catch(Exception e){
 				
 			}
